% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeoPoisson-package.R
\docType{package}
\name{GeoPoisson-package}
\alias{GeoPoisson}
\alias{GeoPoisson-package}
\title{GeoPoisson: GeoPoisson package - estimation via MCMC of parameters and Interpolation map}
\description{
This package estimates the parameter of the Non-homogeneous Poisson model from Morales et al. (2016) package. It also creates an Interpolation map from the observations.
}
\author{
\strong{Maintainer}: Thiago Carneiro \email{thiagomotacarneiro@hotmail.com}

Other contributors:
\itemize{
  \item Fidel Morales \email{Fidel_e_castro@hotmail.com} [contributor]
  \item Antonio Hermes Silva \email{hermes@ccet.ufrn.br} [contributor]
}

}
\keyword{internal}
